refactor: Rename components folder to _components
